Changed the Adafruit code to use RFM69 library
We were using the wrong library, the one we were previously using was
for the RFM95 (the one used on the Adafruit LoRa Feathers). The one we
have though is the RFM69, which is very similar, and nothing really seem
to have changed with the swap. Able to read registers, but not write on
Module B.